Move duplicated PacketFmt classes to common file
