+++ uhal/uhal/include/uhal/ProtocolMmap.hpp
@@ -77,16 +77,6 @@ namespace uhal
   //! Transport protocol to transfer an IPbus buffer via device file, using mmap
   class Mmap : public IPbus< 2 , 0 >
   {
-    public:
-      class PacketFmt {
-      public:
-        PacketFmt(const uint8_t* const, const size_t);
-        PacketFmt(const std::vector< std::pair<const uint8_t*, size_t> >& aData);
-        ~PacketFmt();
-
-        const std::vector< std::pair<const uint8_t*, size_t> > mData;
-      };
-
     private:
       class File {
       public:
