+++ uhal/uhal/include/uhal/ProtocolPCIe.hpp
@@ -81,15 +81,6 @@ namespace uhal
   class PCIe : public IPbus< 2 , 0 >
   {
     public:
-      class PacketFmt {
-      public:
-        PacketFmt(const uint8_t* const, const size_t);
-        PacketFmt(const std::vector< std::pair<const uint8_t*, size_t> >& aData);
-        ~PacketFmt();
-
-        const std::vector< std::pair<const uint8_t*, size_t> > mData;
-      };
-
       class File {
       public:
         File(const std::string& aPath, int aFlags);
@@ -263,8 +254,6 @@ namespace uhal
       //! The list of buffers still awaiting a reply
       std::deque < std::shared_ptr< Buffers > > mReplyQueue;
   };
-
-  std::ostream& operator<<(std::ostream& aStream, const PCIe::PacketFmt& aPacket);
 
 }
