+++ uhal/uhal/src/common/detail/PacketFmt.cpp
@@ -1,0 +1,44 @@
+
+#include "uhal/detail/PacketFmt.hpp"
+
+
+#include <iomanip>
+#include <iostream>
+
+
+namespace uhal {
+namespace detail {
+
+PacketFmt::PacketFmt(const uint8_t* const aPtr, const size_t aNrBytes) :
+  mData(1, std::pair<const uint8_t*, size_t>(aPtr, aNrBytes))
+{}
+
+
+PacketFmt::PacketFmt(const std::vector< std::pair<const uint8_t*, size_t> >& aData) :
+  mData(aData)
+{}
+
+
+PacketFmt::~PacketFmt()
+{}
+
+
+std::ostream& operator<<(std::ostream& aStream, const PacketFmt& aPacket)
+{
+  std::ios::fmtflags lOrigFlags( aStream.flags() );
+
+  size_t lNrBytesWritten = 0;
+  for (size_t i = 0; i < aPacket.mData.size(); i++) {
+    for (const uint8_t* lPtr = aPacket.mData.at(i).first; lPtr != (aPacket.mData.at(i).first + aPacket.mData.at(i).second); lPtr++, lNrBytesWritten++) {
+      if ((lNrBytesWritten & 3) == 0)
+        aStream << std::endl << "   @ " << std::setw(3) << std::dec << (lNrBytesWritten >> 2) << " :  x";
+      aStream << std::setw(2) << std::hex << uint16_t(*lPtr) << " ";
+    }
+  }
+
+  aStream.flags( lOrigFlags );
+  return aStream;
+}
+
+}
+}
